+++ ui/src/app/modules/settings/remove-all-cached-accessories-modal/remove-all-cached-accessories-modal.component.ts
@@ -1,9 +1,8 @@
-import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import { Component } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateService } from '@ngx-translate/core';
 import { ToastrService } from 'ngx-toastr';
-import { ApiService } from '../../../core/api.service';
+import { ApiService } from '@/app/core/api.service';
 
 @Component({
   selector: 'app-remove-all-cached-accessories-modal',