+++ ui/src/app/modules/settings/reset-homebridge-modal/reset-homebridge-modal.component.ts
@@ -3,7 +3,7 @@ import { TranslateService } from '@ngx-translate/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
-import { ApiService } from '../api.service';
+import { ApiService } from '@/app/core/api.service';
 
 @Component({
   templateUrl: './reset-homebridge-modal.component.html',