+++ ui/src/app/modules/settings/settings.module.ts
@@ -6,16 +6,22 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { SettingsRoutingModule } from './settings-routing.module';
 import { SettingsComponent } from './settings.component';
+import { ResetHomebridgeModalComponent } from './reset-homebridge-modal/reset-homebridge-modal.component';
+import { UnpairAccessoryModalComponent } from './unpair-accessory-modal/unpair-accessory-modal.component';
 import { RemoveAllCachedAccessoriesModalComponent } from './remove-all-cached-accessories-modal/remove-all-cached-accessories-modal.component';
 import { RemoveSingleCachedAccessoryModalComponent } from './remove-single-cached-accessory-modal/remove-single-cached-accessory-modal.component';
 
 @NgModule({
   entryComponents: [
+    ResetHomebridgeModalComponent,
+    UnpairAccessoryModalComponent,
     RemoveAllCachedAccessoriesModalComponent,
     RemoveSingleCachedAccessoryModalComponent,
   ],
   declarations: [
     SettingsComponent,
+    ResetHomebridgeModalComponent,
+    UnpairAccessoryModalComponent,
     RemoveAllCachedAccessoriesModalComponent,
     RemoveSingleCachedAccessoryModalComponent,
   ],